gpu bridge: make the shim + protocol portable to 32-bit (WOW64) guests
Building the project for x86 surfaced cross-bitness issues, all now fixed so a
32-bit (WOW64) guest shim can talk to the 64-bit emulator host (and vice versa):

- shim handle<->object_id conversion branched on handle shape: dispatchable
  handles are pointers (32-bit on WOW64) and store the id in the pointer value;
  non-dispatchable handles are uint64_t on every platform and carry the full id.
  The old reinterpret_cast<uintptr_t> truncated the 64-bit id and didn't compile
  for uint64 handles on x86.
- vulkan_shim.def exports the entry points under undecorated names; VKAPI_CALL is
  __stdcall, which would otherwise decorate them on x86 and hide them from the
  guest's GetProcAddress (the real vulkan-1.dll uses a def for the same reason).
- host-side std::vector sizings cast uint64 -> size_t (bounded by 32-bit buffer
  lengths) to silence x86 /WX narrowing.
- static_assert guards pin representative wire-struct sizes so a future pointer-
  sized member fails the build on at least one architecture.

Verified: 32-bit samples + shim run under the 64-bit analyzer's WOW64 path
(vulkan-shim-test fill/clear readback pass; windowed samples render).

+++ src/windows-emulator/devices/gpu_bridge.cpp
@@ -695,7 +695,7 @@ namespace sogen
 
                 const auto copy_bytes = std::min<uint64_t>(request.size, context.output_buffer_length);
 
-                std::vector<std::byte> bytes(copy_bytes);
+                std::vector<std::byte> bytes(static_cast<size_t>(copy_bytes));
                 const int32_t result = this->vulkan_.download_memory(request.device, request.memory, request.offset,
                                                                      copy_bytes, bytes.data(), bytes.size());
                 if (result != 0)
@@ -723,7 +723,7 @@ namespace sogen
 
                 const auto payload = std::min<uint64_t>(request.size, context.input_buffer_length - sizeof(request_t));
 
-                std::vector<std::byte> bytes(payload);
+                std::vector<std::byte> bytes(static_cast<size_t>(payload));
                 if (payload > 0)
                 {
                     win_emu.emu().read_memory(context.input_buffer + sizeof(request_t), bytes.data(), bytes.size());
@@ -958,7 +958,7 @@ namespace sogen
                 }
 
                 const auto code_bytes = std::min<uint64_t>(request.code_size, context.input_buffer_length - sizeof(request_t));
-                std::vector<std::byte> code(code_bytes);
+                std::vector<std::byte> code(static_cast<size_t>(code_bytes));
                 if (code_bytes > 0)
                 {
                     win_emu.emu().read_memory(context.input_buffer + sizeof(request_t), code.data(), code.size());
